extracted all catalog creating functions from catalog_versions.py into new file create_other_catalogs.py
